cuda: rename files to get rid of so many cudas
