+++ cuda/warping_string.go
